extract position, response, do some cleanup, start on pawn movegen
