+++ src/engine/driver/hazel/state.rs
@@ -1,0 +1,7 @@
+#[derive(Clone, Copy, PartialEq, Debug, Default)]
+pub enum State {
+    #[default] Idle,
+    Ready,
+    Pondering,
+    Quitting,
+}
