+++ src/game/chess/familiar.rs
@@ -9,7 +9,6 @@ pub struct Familiar<'a> {
     movesheets: Vec<MoveSheet>,
 }
 
-
 impl<'a> Familiar<'a> {
     pub fn new(cursor: Cursor<'a, Action<Move, BEN>>) -> Self {
         Self { cursor, movesheets: vec![MoveSheet::default()] }
@@ -46,7 +45,6 @@ impl<'a> Familiar<'a> {
                 break;
             }
         }
-
     }
 
     pub fn rewind(&mut self) {
